🚚 Rename package to `mg-substack-csv`
no issue

- This package is far away from being a generic CSV migration tool, but *does* work with Substack, which is also - so far - the only platform that provides export files in this format
