+++ packages/mg-substack-csv/.eslintrc.js
