+++ packages/mg-substack-csv/lib/mapper.js
